Our custom changes
We don't need the bot to be on multiple networks, so remove that code.
We don't need the bot to add/remove/edit IRC Networks for users, so remove that code.
We don't need webhooks so remove that code.
Our own channels and email messages.
Send MODE +B on connecting to the IRC network so the bot is marked as such.